feat(vercel): export express app, add Vercel API handler and rewrites; adjust server entry for local dev
